Refactor topicmap renderer (#1073)
The `TopicmapPanel.vue` component (part of `dm4-topicmaps` module) is gone. Its responsibility was to pass topicmap data to the renderer. Now the (Cytoscape) renderer (part of `dm5-topicmap-panel` module) gets hold of topicmap data on its own by hook-in to the application event flow. It does so by registering its own store module.

Thus the Cytoscape renderer is able to participate in processing WebSocket messages and update the Cytoscape graph model accordingly.