Revise session management (#1080)
In DM4 a HTTP session is created only when a user loggs in.
On logout the session is invalidated.

In DM5 a session is created for every user, regardless if logged in or not.
On logout the session is *not* invalidated.
The logg-state is represented by the existence/absence of the "username" session attribute.

This way DM5's handling of WebSocket connections is more straight forward.
Now a WebSocket connection is *always* associatied with a valid HTTP session.
Note that the DM5 Webclient creates a WebSocket connection at launch time, when no user might logged in.

See #1080