Core API: convenience add() for simple value (#14)
1 new method in `systems.dmx.core.model.ChildTopicsModel`:

    ChildTopicsModel add(String assocDefUri, Object value)

Convenience method to add a *simple* value to a multiple-valued child.
The `value` paramter is the simple value: a String, Integer, Long, Double, or a Boolean.
Primitive values are auto-boxed.

Example: instead of

    childs.add("lotto.number", mf.newTopicModel("lotto.number", new SimpleValue(23)))

you can just write:

    childs.add("lotto.number", 23)

The `TopicModel` is created for you.

Note: for *single-valued* childs a corresponding convenience method (put()) existed already.