Webclient injects `dm5` dependency into plugins (#43)
A plugin's store module can get the `dm5` instance injected.

To do so the plugin's store module must not export the store module directly, but a function that returns the store module. While plugin initialization the Webclient will pass an object with the dependencies to that function. At the moment this object's only property is `dm5`:

```
export default ({dm5}) => ({

  state: {
    ...
  },

  actions: {
    // operate on dm5 here, e.g. dm5.restClient...
  }
})
```

Important: the plugin must not import `dm5`.