Fix HMR in case plugin is deployed (#109)
If the webclient is run in development mode (`npm run dev`) the frontend code of *all* plugins is *not* fetched from server but loaded via HMR. That means while development you have to add your plugin(s) to `modules/dmx-webclient/src/main/js/plugin_manager.js` in order for the frontend code to be loaded.

The browser console remembers you when you're in dev mode.