Core API: typeUri not required for topic/assoc.update() call (#341)
Note: this is NOT a breaking change but a convenience.

The update model you pass to topic/assoc.update() must not necessarily have "typeUri" set.
The value integrator takes the type URI from the topic/assoc then.

This works also if both, topic/assoc retyping and value integration is performed in a single update() call
(in this case the values must be conform to the new type of course).

Before this change the value integrator throwed
"java.lang.IllegalArgumentException: Tried to integrate values whose typeUri is not set"